New OOK encoder, designed by @rascafr
This PR adds a new encoder to the OOK APP.

This encoder works for 8 bit OOK Garage Door LINEAR DELTA 3 DNR00001 DR DR3A Single Channel Gate, as used in the All-O-Matic OH-200 DC Overhead Gate Operator.

This is part of @rascafr work. I took the liberty to generate this PR, so we can add it into the next firmware release, while we still work / test the rest of the proposed changes / additions.