WIP BT LE APP
Still trying to get BLE "scanner" to do something useful.  Had to search for the original code (found it, I placed the url in the comments). It is still showing the wrong Arvertiser address (the "MAC") but at least now the whole thing is understood (and I commented each step). I hope I am close to the real deal