WIP BTLE
For now, this is not solvable for me with the time I got. After the Advertising address, comes the PDU (header / payload). It seems to be "garbage" and the CRC checksum is never checking it. Something is fundamentally wrong. Maybe another pair of eyes can find where is the problem, exactly. I left all my comments on the code, coupled with some commented out testings for inspecting the CRC process.